APS/requirements.txt
requests==2.31.0 
pytest==7.4.0 
pydantic==2.4.2 
pytest-html==4.1.1 
Faker==19.6.2 
python-dotenv==1.0.0 
